pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "notion-cli"
version = "0.1.0"
description = "AI-driven CLI for operating a Notion workspace's databases, cards, blocks, comments and search."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "mat", email = "mat@hummingfood.com" }]
keywords = ["notion", "cli", "ai", "agent", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
]
dependencies = [
    "notion-client>=2.2,<3",
    "typer>=0.12,<1",
    "rich>=13.7",
    "keyring>=24.3",
    "tenacity>=8.2",
    "python-dateutil>=2.8",
    "tomli-w>=1.0",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-mock>=3.12",
    "ruff>=0.5",
    "mypy>=1.10",
    "types-python-dateutil",
    "vcrpy>=6.0",
]

[project.scripts]
notion-cli = "notion_cli.cli:app"

[project.urls]
Homepage = "https://tech.hummingfood.com"

[tool.hatch.build.targets.wheel]
packages = ["src/notion_cli"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers"
markers = [
    "integration: requires NOTION_TEST_TOKEN and NOTION_TEST_DATABASE_ID env vars",
]
